feat: Remove new tr. button from the navigation drawer
