+++ lib/app/onboarding/intro.page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:monekin/app/onboarding/onboarding.dart';
-import 'package:monekin/app/settings/about_page.dart';
+import 'package:monekin/app/settings/widgets/display_app_icon.dart';
 import 'package:monekin/core/presentation/app_colors.dart';
 import 'package:monekin/core/presentation/responsive/breakpoint_container.dart';
 import 'package:monekin/core/presentation/widgets/html_text.dart';
