+++ lib/app/settings/help_us_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:monekin/app/settings/about_page.dart';
 import 'package:monekin/app/settings/purchases/donate_button.dart';
 import 'package:monekin/app/settings/purchases/in_app_purchase.dart';
+import 'package:monekin/app/settings/widgets/display_app_icon.dart';
 import 'package:monekin/app/settings/widgets/setting_card_item.dart';
 import 'package:monekin/core/utils/open_external_url.dart';
 import 'package:monekin/i18n/translations.g.dart';
